requirements.txt
fastapi
uvicorn
pymongo
motor
pydantic
bcrypt
PyJWT>=2.0
python-multipart
reportlab
pandas
python-docx
PyPDF2
python-dotenv
groq
httpx